Add drawer and clean up directories
