+++ src/App.tsx
@@ -7,7 +7,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Launch from './components/LaunchPage/Launch';
 import { QuestionnaireProvider } from './classes/QuestionnaireProvider';
 import LaunchContextProvider from './custom-contexts/LaunchContext';
-import QAuth from './components/QRenderer/QAuth';
+import Auth from './components/Auth';
 import { QuestionnaireResponseProvider } from './classes/QuestionnaireResponseProvider';
 
 const questionnaireProvider = new QuestionnaireProvider();
@@ -30,7 +30,7 @@ function App() {
             <CssBaseline />
             <BrowserRouter>
               <Routes>
-                <Route path="/" element={<QAuth />} />
+                <Route path="/" element={<Auth />} />
                 <Route path="/launch" element={<Launch />} />
               </Routes>
             </BrowserRouter>
