+++ src/components/DebugComponents/DisplayDebugQResponse.tsx
@@ -3,7 +3,7 @@ import { Box, IconButton, Stack, Typography } from '@mui/material';
 import { Bundle, Questionnaire, QuestionnaireResponse } from 'fhir/r5';
 import DeleteIcon from '@mui/icons-material/Delete';
 import ContentCopyIcon from '@mui/icons-material/ContentCopy';
-import { RoundButton } from '../../StyledComponents/Buttons.styles';
+import { RoundButton } from '../StyledComponents/Buttons.styles';
 
 interface Props {
   questionnaire: Questionnaire;
