--- src/components/QRenderer/QDrawer/QOperationsList.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import { Box } from '@mui/material';
-function SideBarOperationList() {
-  return <Box sx={{ my: 1 }}></Box>;
-}
-
-export default SideBarOperationList;
