+++ src/components/QRenderer/QFormComponents/QItemSimple/QItemDecimal.tsx
@@ -7,7 +7,7 @@ import {
 } from '../../../../interfaces/Interfaces';
 import { QuestionnaireItem, QuestionnaireResponseItem } from 'fhir/r5';
 import { createQrItem } from '../../../../functions/QrItemFunctions';
-import { CalcExpressionContext } from '../../QForm';
+import { CalcExpressionContext } from '../../Form';
 
 interface Props
   extends PropsWithQrItemChangeHandler<QuestionnaireResponseItem>,
