--- src/components/QRenderer/QLayout.tsx
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import QForm from './QForm';
-
-function QLayout() {
-  return <QForm />;
-}
-
-export default QLayout;
